backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.1
python-jose==3.3.0
pyjwt==2.8.0
pydantic==2.5.3
requests==2.31.0
python-multipart==0.0.6
